Move files between catalogs.
