+++ CineWaifu.Domain/Model/Color/XyzColor.cs
@@ -1,4 +1,4 @@
-﻿namespace CineWaifu.Domain.Model
+﻿namespace CineWaifu.Domain.Model.Color
 {
     public record class XyzColor(double X, double Y, double Z);
 }
